requirements.txt
aioredis==1.3.1
asgiref~=3.5.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.1
boto3==1.19.3
botocore==1.22.3
certifi==2020.4.5.1
cffi==1.14.6
channels~=4.0.0
channels-redis==3.3.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.6.0
Django~=4.0.8
django-allauth==0.41.0
django-appconf==1.0.4
django-avatar==5.0.0
django-ckeditor~=6.5.1
django-countries==6.1.2
django-crispy-forms==1.9.0
django-debug-toolbar~=3.7.0
django-extensions==2.2.9
django-js-asset==1.2.2
django-mptt~=0.14.0
django-storages==1.12.2
django-thumbs-v2==0.4.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.9
incremental==21.3.0
jmespath==0.10.0
msgpack==1.0.2
msgpack-python==0.5.6
oauthlib==3.1.0
pep8==1.7.1
Pillow~=9.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.14.0
sqlparse==0.3.1
stripe~=5.2.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.25.9
zope.interface==5.4.0
